feat: added popular event handling method

+++ ewm-main-service/src/main/java/ru/yandex/tonychem/ewmmainservice/rating/model/dto/RatingShortDto.java
@@ -5,6 +5,6 @@ import lombok.Value;
 @Value
 public class RatingShortDto {
     Long eventId;
-    Long likeCount;
-    Long dislikeCount;
+    Integer likeCount;
+    Integer dislikeCount;
 }
